change Phones from client component to server
