+++ src/app/(main)/layout.tsx
@@ -1,10 +1,16 @@
+import Phones from '@/components/ui/Phones';
 import Contacts from '@/sections/Contacts';
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <>
       {children}
-      <Contacts />
+      <Contacts>
+        <Phones
+          variant='contacts'
+          className='smOnly:pl-[10px] xl:w-1/2 notXL:mb-[60px]'
+        />
+      </Contacts>
     </>
   );
 };
