+++ src/components/ui/Phones/Phones.types.ts
@@ -2,20 +2,3 @@ export interface IPhonesProps {
   variant: 'contacts' | 'footer';
   className?: string;
 }
-
-export interface IContact {
-  email: string;
-  numbers: {
-    [country: string]: string;
-  };
-}
-
-interface ContactData {
-  attributes: IContact;
-}
-
-export interface IContactResponse {
-  contact: {
-    data: ContactData;
-  };
-}
